Se modificó la interfaz de inicio y ahora no se puede eliminar al usuario con id 1

+++ modules/usuarios/model.php
@@ -4,16 +4,19 @@ require_once "../database.php";
 
 if ($_SESSION['rol_usuario'] == "Admin") {
     if (isset($_POST['users_table'])) {
-        $query_get_users = "SELECT id_usuario, rol_usuario, usuario, nombre_usuario, telefono_usuario, correo_usuario, creacion_cuenta, estado_usuario FROM usuarios";
+        if ($_SESSION['id_usuario'] != 1) {
+            $query_get_users = "SELECT id_usuario, rol_usuario, usuario, nombre_usuario, telefono_usuario, correo_usuario, creacion_cuenta, estado_usuario FROM usuarios WHERE id_usuario != 1";
+        } else {
+            $query_get_users = "SELECT id_usuario, rol_usuario, usuario, nombre_usuario, telefono_usuario, correo_usuario, creacion_cuenta, estado_usuario FROM usuarios";
+        }
+
         $users_data = mysqli_query($conn, $query_get_users);
-    
         $data = mysqli_fetch_all($users_data, MYSQLI_ASSOC);
     
         echo json_encode($data, JSON_UNESCAPED_UNICODE);
     }
 
     if (isset($_POST['action']) && $_POST['action'] == "insert") {
-        $rol_usuario = htmlspecialchars(trim($_POST['rol_usuario']), ENT_QUOTES, 'UTF-8');
         $usuario = htmlspecialchars(trim($_POST['usuario']), ENT_QUOTES, 'UTF-8');
         $contrasenia = htmlspecialchars(trim($_POST['contrasenia']), ENT_QUOTES, 'UTF-8');
         $nombre_usuario = htmlspecialchars(trim($_POST['nombre_usuario']), ENT_QUOTES, 'UTF-8');
@@ -21,10 +24,10 @@ if ($_SESSION['rol_usuario'] == "Admin") {
         $correo_usuario = htmlspecialchars(trim($_POST['correo_usuario']), ENT_QUOTES, 'UTF-8');
         $contrasenia_segura = sha1($contrasenia);
 
-        $query_user_insert = "INSERT INTO usuarios(id_usuario, rol_usuario, usuario, contrasenia, nombre_usuario, telefono_usuario, correo_usuario) VALUES(NULL, '$rol_usuario', '$usuario', '$contrasenia_segura', '$nombre_usuario', '$telefono_usuario', '$correo_usuario')";
+        $query_user_insert = "INSERT INTO usuarios(id_usuario, usuario, contrasenia, nombre_usuario, telefono_usuario, correo_usuario) VALUES(NULL, '$usuario', '$contrasenia_segura', '$nombre_usuario', '$telefono_usuario', '$correo_usuario')";
         $result_user_insert = mysqli_query($conn, $query_user_insert);  
 
-        if ($result_user_insert) echo "Usuario registrado!";
+        if ($result_user_insert) echo json_encode(["icon" => "success", "msg" => "Usuario registrado!"], JSON_UNESCAPED_UNICODE);
     }
 
     if (isset($_POST['edit_id'])) {
@@ -39,7 +42,6 @@ if ($_SESSION['rol_usuario'] == "Admin") {
 
     if (isset($_POST['action']) && $_POST['action'] == "update") {
         $id_usuario = htmlspecialchars(trim($_POST['id_usuario']), ENT_QUOTES, 'UTF-8');
-        $rol_usuario = htmlspecialchars(trim($_POST['rol_usuario']), ENT_QUOTES, 'UTF-8');
         $usuario = htmlspecialchars(trim($_POST['usuario']), ENT_QUOTES, 'UTF-8');
         $contrasenia = htmlspecialchars(trim($_POST['contrasenia']), ENT_QUOTES, 'UTF-8');
         $nombre_usuario = htmlspecialchars(trim($_POST['nombre_usuario']), ENT_QUOTES, 'UTF-8');
@@ -49,24 +51,29 @@ if ($_SESSION['rol_usuario'] == "Admin") {
         if ($_SESSION['id_usuario'] == $id_usuario) $_SESSION['nombre_usuario'] = $nombre_usuario;
         
         if ($contrasenia == "") {
-            $query_user_update = "UPDATE usuarios SET rol_usuario = '$rol_usuario', usuario = '$usuario', nombre_usuario = '$nombre_usuario', telefono_usuario = '$telefono_usuario', correo_usuario = '$correo_usuario', estado_usuario = '$estado_usuario' WHERE id_usuario = $id_usuario";
+            $query_user_update = "UPDATE usuarios SET usuario = '$usuario', nombre_usuario = '$nombre_usuario', telefono_usuario = '$telefono_usuario', correo_usuario = '$correo_usuario', estado_usuario = '$estado_usuario' WHERE id_usuario = $id_usuario";
         } else {
             $contrasenia_segura = sha1($contrasenia);
-            $query_user_update = "UPDATE usuarios SET rol_usuario = '$rol_usuario', usuario = '$usuario', contrasenia = '$contrasenia_segura', nombre_usuario = '$nombre_usuario', telefono_usuario = '$telefono_usuario', correo_usuario = '$correo_usuario', estado_usuario = '$estado_usuario' WHERE id_usuario = $id_usuario";
+            $query_user_update = "UPDATE usuarios SET usuario = '$usuario', contrasenia = '$contrasenia_segura', nombre_usuario = '$nombre_usuario', telefono_usuario = '$telefono_usuario', correo_usuario = '$correo_usuario', estado_usuario = '$estado_usuario' WHERE id_usuario = $id_usuario";
         }
 
         $result_user_update = mysqli_query($conn, $query_user_update);
         
-        if ($result_user_update) echo "Datos del usuario actualizados!";
+        if ($result_user_update) echo json_encode(["icon" => "success", "msg" => "Datos del usuario actualizados!"], JSON_UNESCAPED_UNICODE);
     }
 
     if (isset($_POST['delete_id'])) {
         $id_usuario = $_POST['delete_id'];
                 
+        if ($_SESSION['id_usuario'] == $id_usuario) {
+            echo json_encode(["icon" => "error", "msg" => "No puedes eliminar tu propio usuario!"], JSON_UNESCAPED_UNICODE);
+            exit;
+        }
+
         $query_user_delete = "DELETE FROM usuarios WHERE id_usuario = $id_usuario";
         $result_user_delete = mysqli_query($conn, $query_user_delete);    
 
-        if ($result_user_delete) echo "Usuario eliminado!";
+        if ($result_user_delete) echo json_encode(["icon" => "success", "msg" => "Usuario eliminado!"], JSON_UNESCAPED_UNICODE);
     }
 }
